Moving prove and verify in algorithm.rs
